ewaste-classifier/backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
python-multipart
numpy
